+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Addition of Binary Actors (supported openHAB item: `Switch`):
     * Switch
     * Fan
+    * Outlet (with optional Outlet In Use Characteristic)
 
 ## V0.5.0
 * Addition of CurrentPosition/TargetPosition Actors (Supported openHAB items: `Rollershutter`, `Number` and `Switch`) with optional item representing the service's state (Supported openHAB items: `Rollershutter`, `Number`, `Switch`, `Contact`)